panecho/app/requirements.txt
numpy==1.26.4
pandas==2.2.3
timm==1.0.20
pillow==11.1.0
